fix: handle case displayThankYouMessage is false
