+++ PostHog/Surveys/SurveysRootView.swift
@@ -20,10 +20,8 @@
                         survey: survey,
                         isSurveyCompleted: displayManager.isSurveyCompleted,
                         currentQuestionIndex: displayManager.currentQuestionIndex ?? 0,
-                        onClose: displayManager.userDismissedSurvey,
-                        onNextQuestionClicked: { index, response in
-                            displayManager.onNextQuestion(index: index, response: response)
-                        }
+                        onClose: displayManager.dismissSurvey,
+                        onNextQuestionClicked: displayManager.onNextQuestion
                     )
                     .environment(\.colorScheme, .light) // enforce light theme for now
                 }
@@ -35,8 +33,9 @@
                     displayManager.displayedSurvey
                 },
                 set: { newValue in
+                    // in case interactive dismiss is allowed
                     if newValue == nil {
-                        displayManager.hideSurvey()
+                        displayManager.dismissSurvey()
                     }
                 }
             )
